+++ ast_canopy/CMakeLists.txt
@@ -14,7 +14,7 @@ project(${SKBUILD_PROJECT_NAME}
 
 find_package(Python REQUIRED COMPONENTS Interpreter Development.Module)
 find_package(pybind11 CONFIG REQUIRED)
-find_package(astcanopy 0.1.0 REQUIRED)
+find_package(astcanopy REQUIRED HINTS /usr/local/lib/cmake)
 
 python_add_library(pylibastcanopy MODULE ast_canopy/pylibastcanopy.cpp WITH_SOABI)
 target_link_libraries(pylibastcanopy PRIVATE pybind11::headers)
