+++ ast_canopy/cpp/src/record.cpp
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // clang-format on
 
-#include "ast_canopy.hpp"
+#include <ast_canopy/ast_canopy.hpp>
 
 #include <clang/AST/ASTContext.h>
 #include <clang/AST/DeclCXX.h>
